+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* Register .tsx for TypeScript (#315, backus)
+
 ## 2.0.7 (2017-03-19)
 
 * Do not modify BasicObject during template compilation on ruby 2.0+ (#309, jeremyevans)